Reorganized classes in Rest.?pp by moving Response and CurlRest to Response.?pp and CurlRest.?pp respectively. Modified CMakeLists.txt, main.cpp, and rest_test.cpp to facilitate these changes.
